broke the graph tab and search functionality out into their own classes
